folder structure update, movie list fixes
