+++ frontend/src/pages/admin/UserList.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import api from "../utils/api";
+import api from "../../utils/api";
 
 const UserList = () => {
   const [users, setUsers] = useState([]);
