+++ frontend/src/pages/user/Login.js
@@ -1,7 +1,7 @@
 import React, { useState, useContext } from "react";
 import { useNavigate } from "react-router-dom";
-import { UserContext } from "../context/UserContext";
-import api from "../utils/api";
+import { UserContext } from "../../context/UserContext";
+import api from "../../utils/api";
 
 const Login = () => {
     const [username, setUsername] = useState("");
