+++ frontend/src/pages/user/Logout.js
@@ -1,6 +1,6 @@
 import React, { useContext, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import { UserContext } from "../context/UserContext";
+import { UserContext } from "../../context/UserContext";
 
 const Logout = () => {
   const { logout } = useContext(UserContext); // Use logout from UserContext
