+++ frontend/src/pages/user/Register.js
@@ -1,33 +1,28 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import api from "../utils/api";
-import UserForm from "../components/UserForm";
+import api from "../../utils/api";
+import UserForm from "../../components/UserForm";
 
-const RegisterAdmin = () => {
+const Register = () => {
   const [formData, setFormData] = useState({
     username: "",
     first_name: "",
     last_name: "",
     email: "",
     password: "",
-    role: "", // Default role
   });
   const navigate = useNavigate();
 
   const handleInputChange = (e) => {
     const { name, value } = e.target;
-    setFormData({ ...formData, [name]: value || "" }); // Ensure fallback to empty string
+    setFormData({ ...formData, [name]: value });
   };
 
   const handleRegister = async () => {
     try {
-      const token = localStorage.getItem("token");
-      await api.post("/users/register-admin", formData, {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-      alert(`Registration of ${formData.username} successful!`);
+      await api.post("/users/register", formData);
+      alert("Registration successful! Please log in.");
+      navigate("/login");
     } catch (error) {
       console.error("Registration error:", error.response || error.message);
       alert(error.response?.data?.detail || "Failed to register.");
@@ -36,7 +31,7 @@ const RegisterAdmin = () => {
 
   return (
     <div className="container mt-5">
-      <h1 className="text-center mb-4">Register Admin</h1>
+      <h1 className="text-center mb-4">Register</h1>
       <div className="row justify-content-center">
         <div className="col-md-6">
           <div className="card p-4 shadow-sm">
@@ -44,7 +39,6 @@ const RegisterAdmin = () => {
               <UserForm
                 formData={formData}
                 handleInputChange={handleInputChange}
-                includeRole={true}
               />
               <button
                 className="btn btn-primary w-100"
@@ -60,4 +54,4 @@ const RegisterAdmin = () => {
   );
 };
 
-export default RegisterAdmin;
+export default Register;
